+++ server/routes/mycelium.js
@@ -195,6 +195,7 @@ import { registerTeamRoutes } from './teams.js';
 import { registerAssetRoutes } from './assets.js';
 import { registerProjectRoutes } from './projects.js';
 import { registerConceptRoutes } from './concepts.js';
+import { registerSkillRoutes } from './skills.js';
 
 var ADMIN_KEY = process.env.ADMIN_KEY;
 function isAdminKey(key) {
@@ -1864,72 +1865,6 @@ function parseVoiceCommand(text, who) {
     response: 'I heard: "' + text + '". Try asking about agent status, tasks, bugs, or drones.'
   };
 }
-
-// ======== SKILLS REGISTRY ========
-
-router.get('/skills', asyncHandler(function (req, res) {
-  var who = checkAgentOrAdmin(req, res);
-  if (!who) return;
-  var skills = listSkills({
-    category: req.query.category,
-    search: req.query.search
-  });
-  res.json(skills);
-}));
-
-router.get('/skills/:id', asyncHandler(function (req, res) {
-  var who = checkAgentOrAdmin(req, res);
-  if (!who) return;
-  var skill = getSkill(req.params.id);
-  if (!skill) return res.status(404).json({ error: 'skill not found' });
-  res.json(skill);
-}));
-
-router.post('/skills', asyncHandler(function (req, res) {
-  var who = checkAdmin(req, res);
-  if (!who) return;
-  var b = req.body;
-  if (!b.id || !b.name) return res.status(400).json({ error: 'id and name required' });
-  try {
-    var result = createSkill(b.id, b.name, b.description, b.category, b.version, b.author,
-      b.install_type, b.install_data, b.required_capabilities, b.tags);
-    emitEvent('skill_created', 'admin', '', b.name, { skill_id: b.id });
-    res.status(201).json(result);
-  } catch (err) {
-    if (err.message && err.message.includes('UNIQUE')) return res.status(409).json({ error: 'skill already exists' });
-    throw err;
-  }
-}));
-
-router.put('/skills/:id', asyncHandler(function (req, res) {
-  var who = checkAdmin(req, res);
-  if (!who) return;
-  var updated = updateSkill(req.params.id, req.body);
-  if (!updated) return res.status(404).json({ error: 'skill not found' });
-  res.json(updated);
-}));
-
-// Agent skill management
-router.post('/skills/:id/install', asyncHandler(function (req, res) {
-  var who = checkAgentOrAdmin(req, res);
-  if (!who) return;
-  var agentId = (who === '__admin__' || who === '__system__') ? (req.body.agent_id || who) : who;
-  if (!agentId) return res.status(400).json({ error: 'agent_id required' });
-  var skill = getSkill(req.params.id);
-  if (!skill) return res.status(404).json({ error: 'skill not found' });
-  installSkill(agentId, req.params.id, req.body.config);
-  emitEvent('skill_installed', agentId, '', skill.name, { skill_id: req.params.id });
-  res.json({ ok: true, skill_id: req.params.id, agent_id: agentId });
-}));
-
-router.post('/skills/:id/uninstall', asyncHandler(function (req, res) {
-  var who = checkAgentOrAdmin(req, res);
-  if (!who) return;
-  var agentId = (who === '__admin__' || who === '__system__') ? (req.body.agent_id || who) : who;
-  if (!agentId) return res.status(400).json({ error: 'agent_id required' });
-  uninstallSkill(agentId, req.params.id);
-  res.json({ ok: true });
-}));
 
 router.get('/agents/:agentId/skills', asyncHandler(function (req, res) {
   var who = checkAgentOrAdmin(req, res);
@@ -3755,6 +3690,10 @@ registerDroneRoutes(router, {
   apiError, emitEvent, getAdminDisplayName, isAdminKey, getStudioUser,
   requireAuth, artifactUpload, ARTIFACTS_DIR,
   DRONE_JOB_STATUSES,
+});
+
+registerSkillRoutes(router, {
+  asyncHandler, checkAgentOrAdmin, checkAdmin, emitEvent,
 });
 
 registerApprovalRoutes(router, {
